rename submission to Transcribe! move footer, capitalize file names,
