+++ frontend/src/components/Landing/Landing.jsx
@@ -2,7 +2,7 @@ import React, { useEffect } from "react";
 import Introduction from "./Introduction/Introduction";
 import Features from "./Features/Features";
 import Divider from "../Common/Divider";
-import Footer from "../Footer/Footer";
+import Footer from "../Common/Footer/Footer";
 import About from "./About/About";
 
 export default function Landing() {
