+++ frontend/src/components/Main/Transcribe/CollapsedTimeline.jsx
