+++ frontend/src/components/Main/Transcribe/FullTranscript.jsx
