+++ frontend/src/components/Main/Transcribe/QuestionCategoryDistribution.jsx
