+++ frontend/src/components/Main/Transcribe/TalkingDistribution.jsx
