+++ frontend/src/components/Main/Transcribe/WordCloud.jsx
