+++ frontend/src/components/Main/Transcribe/transcribe.css
