+++ frontend/src/components/Main/Transcribe/transcript.scss
