+++ frontend/src/components/Main/mainUserPage.jsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { Routes, Route, Outlet } from "react-router-dom";
 import SideMenu from "./SideMenu";
-import Submission from "./Submission/Submission";
+import Transcribe from "./Transcribe/Transcribe";
 import AllFiles from "./AllFiles/AllFiles" ; // hypothetical component
-import Account from "../Account/accountPage";
+import Account from "../Account/AccountPage";
 import "./MainPage.css";
 
 export const Main = () => {
@@ -12,7 +12,7 @@ export const Main = () => {
             <SideMenu />
             <div className="main-content">
                 <Routes>
-                    <Route path="transcribe" element={<Submission />} /> 
+                    <Route path="transcribe" element={<Transcribe />} /> 
                     <Route path="files" element={<AllFiles />} /> 
                     <Route path="account" element={<Account />} />
                     {/* other nested routes if needed */}
